+++ src/index.ts
@@ -1,5 +1,10 @@
 export { parseMd, revertMdAstNode, revertMdAstNode as stringifyMdAst } from './parse-md';
 export { parseMdWithSourceMap } from './source-map/build-source-map';
+export {
+  SourceMapError,
+  SourceMapConsistencyError,
+  SourceMapUnavailableError,
+} from './source-map/errors';
 
 export * from './types';
 export * from './source-map/types';